create new endpoint - post (create new book & books coll.)

+++ models/book.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => {
     {
       title: DataTypes.STRING,
       author: DataTypes.STRING,
-      ISBN: DataTypes.INTEGER,
+      ISBN: DataTypes.BIGINT,
       category: DataTypes.STRING,
       description: DataTypes.STRING(65535),
       image: DataTypes.STRING,
